requirements.txt
Django==1.8.8
argparse==1.4.0
beautifulsoup4==4.4.1
django-debug-toolbar==1.4
django-extensions==1.6.1
django-grappelli==2.7.3
django-htmlmin==0.8.0
django-localflavor==1.2
django-reversion==1.10.0
django-robots==1.1
gunicorn==19.4.5
html5lib==1.0b8
six==1.10.0
sqlparse==0.1.18
wsgiref==0.1.2
